Refactor login to separate route
