+++ app/app.js
@@ -7,6 +7,7 @@ var bodyParser = require('body-parser');
 
 var index = require('./routes/index');
 var register = require('./routes/register');
+var login = require('./routes/login');
 var users = require('./routes/users');
 
 var app = express();
@@ -29,6 +30,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', index);
 app.use('/register', register);
+app.use('/login', login);
 app.use('/users', users);
 
 // catch 404 and forward to error handler
